requirements.txt
streamlit
PyPDF2
pikepdf
reportlab
Pillow
pycryptodome
PyMuPDF
python-docx
weasyprint
openpyxl
python-pptx
pdfplumber
fitz
pytesseract
easyocr
torch
numpy
fastapi==0.104.1
uvicorn==0.24.0
httpx==0.25.2
python-multipart==0.0.6